fixed mongoservererror ns not found

+++ migrations/create-teams.mjs
@@ -9,49 +9,51 @@ const data = JSON.parse(fs.readFileSync(dataPath, 'utf8'));
 const tracks = [...new Set(data.tracks)];
 
 export async function up(db) {
-  await db.createCollection('teams', {
-    validator: {
-      $jsonSchema: {
-        bsonType: 'object',
-        title: 'Teams Object Validation',
-        required: ['teamNumber', 'tableNumber', 'name', 'tracks', 'active'],
-        properties: {
-          _id: {
-            bsonType: 'objectId',
-            description: '_id must be an ObjectId',
-          },
-          teamNumber: {
-            bsonType: 'int',
-            description: 'teamNumber must be an integer',
-          },
-          tableNumber: {
-            bsonType: 'int',
-            description: 'tableNumber must be an integer',
-          },
-          name: {
-            bsonType: 'string',
-            description: 'name must be a string',
-          },
-          tracks: {
-            bsonType: 'array',
-            maxItems: 6,
-            items: {
-              enum: tracks,
-              description: 'track must be one of the valid tracks',
+  if (!(await db.listCollections({ name: 'teams' }).hasNext()))
+    await db.createCollection('teams', {
+      validator: {
+        $jsonSchema: {
+          bsonType: 'object',
+          title: 'Teams Object Validation',
+          required: ['teamNumber', 'tableNumber', 'name', 'tracks', 'active'],
+          properties: {
+            _id: {
+              bsonType: 'objectId',
+              description: '_id must be an ObjectId',
             },
-            description: 'tracks must be an array of strings',
+            teamNumber: {
+              bsonType: 'int',
+              description: 'teamNumber must be an integer',
+            },
+            tableNumber: {
+              bsonType: 'int',
+              description: 'tableNumber must be an integer',
+            },
+            name: {
+              bsonType: 'string',
+              description: 'name must be a string',
+            },
+            tracks: {
+              bsonType: 'array',
+              maxItems: 6,
+              items: {
+                enum: tracks,
+                description: 'track must be one of the valid tracks',
+              },
+              description: 'tracks must be an array of strings',
+            },
+            active: {
+              bsonType: 'bool',
+              description: 'active must be a boolean',
+            },
           },
-          active: {
-            bsonType: 'bool',
-            description: 'active must be a boolean',
-          },
+          additionalProperties: false,
         },
-        additionalProperties: false,
       },
-    },
-  });
+    });
 }
 
 export async function down(db) {
-  await db.collection('teams').drop();
+  if (await db.listCollections({ name: 'teams' }).hasNext())
+    await db.collection('teams').drop();
 }
